Chore: Abstract out closely related code as classes
